Add a bunch more documentation (and rewrite README)
In this commit, I add a bunch more documentation of how to use
genqlient, including a getting started guide, and a FAQ that also acts
as a how-to shaped index of common configuration options and extension
points.  Much of this is adapted from the internal doc I wrote at Khan,
or reorganized from what was in the README.  Speaking of which, now that
there are better places for all the details, I rewrote the README to be
a bit more of an overview and index of other documentation.  (For now I
left the "unmaintained" notes, which I'll remove fairly soon once all
our ducks are in a row.)

Fixes #26.

Issue: https://github.com/Khan/genqlient/issues/26

Test plan:
make check, read the docs

Reviewers: marksandstrom, adam, miguel, mahtab, jvoll