Include __typename on interface-typed fragments
We ask for `__typename` on interface-typed fields so that we know which
type to use when unmarshaling.  We never strictly need this for
fragments, because the context into which they're spread must either
have `__typename` or must be of concrete type.  But a few cases have
come up where it would be somewhat convenient: first, if you're using
genqlient's types for mocking (not really supported right now, but
possible); and second, for consistency if we add an option to "flatten"
fragments (#30), which would effectively drop the `__typename` field
from the structs otherwise.  Neither is a deeply compelling reason, I
feel, but it's not too hard to do so we may as well.

Test plan:
make check

Reviewers: marksandstrom, steve, jvoll, adam, miguel, mahtab