Remove security issue template
I figured we should have this to say "STOP! Use HackerOne instead".
But GitHub anticipated that and already links to the security policy
(`docs/SECURITY.md`) as an option when [creating an issue][1], so
it's just duplicative.  In this commit I remove it.

[1]: https://github.com/Khan/genqlient/issues/new/choose

Test plan:
above link should show just one option for security

Auditors: steve