Release v0.2.0
With the blog post going up soon, and a few new features, it's time for
another release!  (Plus, Craig has run the webapp tests on this commit,
so it's now extra-tested.)  This commit updates the changelog, and I'll
tag it with the release once it lands.

Test plan:
crossed fingers

Auditors: steve, marksandstrom, csilvers