Add white background for social preview image
This has the same dark mode problem (if your Twitter client, say, is in
dark mode), but we can't be so fancy with CSS, I uess.  So I just
switched to a white background, which looks fine in context.  The actual
source of truth for the social preview image is in the github settings
page, but I checked in the image for reference.

Test plan:
looks fine in the compose window on Facebook

Auditors: steve, marksandstrom